chore(core): rename viewsets/ticket_comment.py -> viewsets/ticket_comment_depreciated.py
ref: #732 #726
